Free NSFW Python DiscordBot/requirements.txt
discord.py
json 
asyncio